--- src/main/java/io/github/eappezo/soundary/advice/common/configuration/FirebaseConfig.java
@@ -1,29 +1,0 @@
-package io.github.eappezo.soundary.advice.common.configuration;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import jakarta.annotation.PostConstruct;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-
-import java.io.InputStream;
-
-@Configuration
-public class FirebaseConfig {
-    @PostConstruct
-    public void init(){
-        try{
-            InputStream serviceAccount = new ClassPathResource("soundaryFirebaseKey.json").getInputStream();
-            FirebaseOptions options = new FirebaseOptions.Builder()
-                    .setCredentials(GoogleCredentials.fromStream(serviceAccount))
-                    .build();
-
-            if (FirebaseApp.getApps().isEmpty()) { // FirebaseApp이 이미 초기화되어 있지 않은 경우에만 초기화 실행
-                FirebaseApp.initializeApp(options);
-            }
-        } catch (Exception e){
-            e.printStackTrace();
-        }
-    }
-}
