+++ src/main/java/io/github/eappezo/soundary/advice/identification/IdentifierConverter.java
@@ -1,4 +1,4 @@
-package io.github.eappezo.soundary.advice;
+package io.github.eappezo.soundary.advice.identification;
 
 import io.github.eappezo.soundary.core.identification.Identifier;
 import org.springframework.core.convert.converter.Converter;
